get nft default to work
