+++ libs/community-contract/config.json
@@ -1,3 +1,3 @@
 {
-  "contractName": "dev-1619398846158-6592600"
+  "contractName": "dev-1623353929808-71316364508823"
 }
