+++ libs/community-contract/contract/dev-deploy.sh
@@ -1,3 +1,4 @@
 #!/bin/sh
-near dev-deploy target/wasm32-unknown-unknown/release/baf_wallet_pk_helper.wasm
-near call dev-1619398846158-6592600 new --accountId=levtester.testnet
+near dev-deploy target/wasm32-unknown-unknown/release/community_contract.wasm | echo
+near call dev-1623353929808-71316364508823 new --accountId=levtester.testnet
+near call dev-1623353929808-71316364508823 add_admins --accountId=levtester.testnet --args '{"new_admins": ["levl333.testnet"]}'
