+++ libs/community-contract/contract/neardev/dev-account
@@ -1,1 +1,1 @@
-dev-1619398846158-6592600
+dev-1623353929808-71316364508823
