+++ libs/community-contract/contract/neardev/dev-account.env
@@ -1,1 +1,1 @@
-CONTRACT_NAME=dev-1619398846158-6592600
+CONTRACT_NAME=dev-1623353929808-71316364508823
