+++ libs/community-contract/contract/src/nft.rs
@@ -1,11 +1,11 @@
-use crate::errors::throw_error;
 use crate::env::predecessor_account_id;
+use crate::errors::throw_error;
 use std::convert::TryInto;
 
-use near_sdk::{
-    env::{current_account_id, is_valid_account_id, keccak256, signer_account_id},
-    near_bindgen, AccountId,
-};
+use near_sdk::env::{current_account_id, is_valid_account_id, keccak256, signer_account_id};
+use near_sdk::AccountId;
+use near_sdk::PanicOnDefault;
+use near_sdk::{collections::UnorderedMap, env, near_bindgen};
 
 use crate::{AccountInfo, Community, CommunityContract, SecpPK, SecpPKInternal};
 
@@ -14,19 +14,6 @@ use crate::{AccountInfo, Community, CommunityContract, SecpPK, SecpPKInternal};
 pub trait NFTFunc {
     fn set_default_nft_contract(&mut self, nft_contract: AccountId);
     fn get_default_nft_contract(&self) -> &Option<AccountId>;
-}
-
-#[near_bindgen]
-impl NFTFunc for Community {
-    fn set_default_nft_contract(&mut self, nft_contract: AccountId) {
-        if !self.admins.contains(&predecessor_account_id()) {
-            throw_error(crate::errors::UNAUTHORIZED);
-        }
-        self.default_nft_contract = Some(nft_contract);
-    }
-    fn get_default_nft_contract(&self) -> &Option<AccountId> {
-        &self.default_nft_contract
-    }
 }
 
 #[cfg(not(target_arch = "wasm32"))]
